[spinor]add spi nor flash support

+++ spinand.c
@@ -1,16 +1,16 @@
 #include <spinand.h>
 
-int spinand_detect(struct xfel_ctx_t * ctx)
+uint64_t spinand_detect(struct xfel_ctx_t * ctx)
 {
 	return 0;
 }
 
-size_t spinand_read(struct xfel_ctx_t * ctx, uint32_t addr, void * buf, size_t len)
+int spinand_read(struct xfel_ctx_t * ctx, uint64_t addr, void * buf, uint64_t len)
 {
 	return 0;
 }
 
-size_t spinand_write(struct xfel_ctx_t * ctx, uint32_t addr, void * buf, size_t len)
+int spinand_write(struct xfel_ctx_t * ctx, uint64_t addr, void * buf, uint64_t len)
 {
 	return 0;
 }
